perf(bundle): tarski-data AXIOM_LIBRARY off eager bundle + TimeDial lazy
Two largest remaining critical-path imports finally split out.

1. tarski-data split:
   - New tarski-flags.ts (76 LOC) houses applyTarskiFlags +
     clearTarskiFlags (both pure graph transforms) and the
     TarskiValidationReport type. None of these touches AXIOM_LIBRARY.
   - tarski-data.ts re-exports these for backward compat but no
     longer defines them locally.
   - useApexStore imports the lightweight trio from tarski-flags
     eagerly. runTarskiValidation (which carries the 891-LOC
     AXIOM_LIBRARY + 800-LOC validation engine) is dynamic-imported
     via a new loadRunTarskiValidation() helper.
   - 4 call sites converted: runTarskiWithAxioms + setTruthFilter
     wrap a simple load->then->set. applyFeedBatch + promoteAutoBridge
     split into two phases — sync mutation + omega adjustments first,
     async tarski revalidation after (gated on s.truthFilter still
     being 'verified' at fire-time).
   - Trade-off: verified-mode feed events show one render of
     un-flagged state before tarski revalidates (microtask hop, cached
     after first load).

2. TimeDial (1207 LOC) is the largest single static-imported component
   on the critical path. Converted to next/dynamic with a placeholder
   matching the dial's height + border style so the layout doesn't
   jump when the chunk lands.

vitest 1524/1524 pass. tsc clean.
